SDK-334 change observation to hold all the things

+++ cpp_sdk/cpp_dist_bundle/benchmarks/memory_high_water_mark.h
@@ -8,11 +8,11 @@ public:
     MemoryHighWaterMarkTracker();
 
     void resetVmHighWaterMark();
-    double getVmHighWaterMark() const;
-    double getDifferenceFromBaseline() const;
+    float getVmHighWaterMark() const;
+    float getDifferenceFromBaseline() const;
 
 private:
-    double m_baseline;
+    float m_baseline;
 };
 
 } // namespace Benchmarks
